Element Operators

Element Operators

Canary.baryweights — Function.

baryweights(r)

returns the barycentric weights associated with the array of points r

Reference: Jean-Paul Berrut & Lloyd N. Trefethen, "Barycentric Lagrange Interpolation", SIAM Review 46 (2004), pp. 501-517. https://doi.org/10.1137/S0036144502417715

source

Canary.interpolationmatrix — Function.

interpolationmatrix(rsrc::AbstractVector{T}, rdst::AbstractVector{T},
                    wbsrc=baryweights(rsrc)::AbstractVector{T}) where T

returns the polynomial interpolation matrix for interpolating between the points rsrc (with associated barycentric weights wbsrc) and rdst

Reference: Jean-Paul Berrut & Lloyd N. Trefethen, "Barycentric Lagrange Interpolation", SIAM Review 46 (2004), pp. 501-517. https://doi.org/10.1137/S0036144502417715

source

Canary.lglpoints — Function.

lglpoints(::Type{T}, N::Integer) where T <: AbstractFloat

returns the points r and weights w associated with the N+1-point Gauss-Legendre-Lobatto quadrature rule of type T

source

Canary.lgpoints — Function.

lgpoints(::Type{T}, N::Integer) where T <: AbstractFloat

returns the points r and weights w associated with the N+1-point Gauss-Legendre quadrature rule of type T

source

Canary.spectralderivative — Function.

spectralderivative(r::AbstractVector{T},
                   wb=baryweights(r)::AbstractVector{T}) where T

returns the spectral differentiation matrix for a polynomial defined on the points r with associated barycentric weights wb

Reference: Jean-Paul Berrut & Lloyd N. Trefethen, "Barycentric Lagrange Interpolation", SIAM Review 46 (2004), pp. 501-517. https://doi.org/10.1137/S0036144502417715

source