Metric Terms

Metric Terms

Canary.creategrid — Function.

creategrid(::Val{1}, elemtocoord::AbstractArray{S, 3},
           r::AbstractVector{T}) where {S, T}

Create a grid using elemtocoord (see brickmesh) using the 1-D (-1, 1) reference coordinates r. The element grids are filled using bilinear interpolation of the element coordinates.

The grid is returned as a tuple of with x array

source

creategrid(::Val{2}, elemtocoord::AbstractArray{S, 3},
           r::AbstractVector{T}) where {S, T}

Create a 2-D tensor product grid using elemtocoord (see brickmesh) using the 1-D (-1, 1) reference coordinates r. The element grids are filled using bilinear interpolation of the element coordinates.

The grid is returned as a tuple of the x and y arrays

source

creategrid(::Val{3}, elemtocoord::AbstractArray{S, 3},
           r::AbstractVector{T}) where {S, T}

Create a 3-D tensor product grid using elemtocoord (see brickmesh) using the 1-D (-1, 1) reference coordinates r. The element grids are filled using bilinear interpolation of the element coordinates.

The grid is returned as a tuple of the x, y, z arrays

source

Canary.creategrid! — Function.

creategrid!(x, elemtocoord, r)

Create a 1-D grid using elemtocoord (see brickmesh) using the 1-D (-1, 1) reference coordinates r. The element grids are filled using linear interpolation of the element coordinates.

If Nq = length(r) and nelem = size(elemtocoord, 3) then the preallocated array x should be Nq * nelem == length(x).

source

creategrid!(x, y, elemtocoord, r)

Create a 2-D tensor product grid using elemtocoord (see brickmesh) using the 1-D (-1, 1) reference coordinates r. The element grids are filled using bilinear interpolation of the element coordinates.

If Nq = length(r) and nelem = size(elemtocoord, 3) then the preallocated arrays x and y should be Nq^2 * nelem == size(x) == size(y).

source

creategrid!(x, y, z, elemtocoord, r)

Create a 3-D tensor product grid using elemtocoord (see brickmesh) using the 1-D (-1, 1) reference coordinates r. The element grids are filled using trilinear interpolation of the element coordinates.

If Nq = length(r) and nelem = size(elemtocoord, 3) then the preallocated arrays x, y, and z should be Nq^3 * nelem == size(x) == size(y) == size(z).

source

Canary.computemetric — Function.

computemetric(x::AbstractArray{T, 2}, D::AbstractMatrix{T}) where T

Compute the 1-D metric terms from the element grid array x using the derivative matrix D. The derivative matrix D should be consistent with the reference grid r used in creategrid!.

The metric terms are returned as a 'NamedTuple` of the following arrays:

J the Jacobian determinant

ξx derivative ∂r / ∂x'

sJ the surface Jacobian

'nx` outward pointing unit normal in $x$-direction

source

computemetric(x::AbstractArray{T, 3}, y::AbstractArray{T, 3},
              D::AbstractMatrix{T}) where T

Compute the 2-D metric terms from the element grid arrays x and y using the derivative matrix D. The derivative matrix D should be consistent with the reference grid r used in creategrid!.

The metric terms are returned as a 'NamedTuple` of the following arrays:

J the Jacobian determinant

ξx derivative ∂r / ∂x'

ηx derivative ∂s / ∂x'

ξy derivative ∂r / ∂y'

ηy derivative ∂s / ∂y'

sJ the surface Jacobian

'nx` outward pointing unit normal in $x$-direction

'ny` outward pointing unit normal in $y$-direction

source

computemetric(x::AbstractArray{T, 3}, y::AbstractArray{T, 3},
              D::AbstractMatrix{T}) where T

Compute the 3-D metric terms from the element grid arrays x, y, and z using the derivative matrix D. The derivative matrix D should be consistent with the reference grid r used in creategrid!.

The metric terms are returned as a 'NamedTuple` of the following arrays:

J the Jacobian determinant

ξx derivative ∂r / ∂x'

ηx derivative ∂s / ∂x'

ζx derivative ∂t / ∂x'

ξy derivative ∂r / ∂y'

ηy derivative ∂s / ∂y'

ζy derivative ∂t / ∂y'

ξz derivative ∂r / ∂z'

ηz derivative ∂s / ∂z'

ζz derivative ∂t / ∂z'

sJ the surface Jacobian

'nx` outward pointing unit normal in $x$-direction

'ny` outward pointing unit normal in $y$-direction

'nz` outward pointing unit normal in $z$-direction

Note

The storage of the volume terms and surface terms from this function are slightly different. The volume terms used Cartesian indexing whereas the surface terms use linear indexing.

source

Canary.computemetric! — Function.

computemetric!(x, J, ξx, sJ, nx, D)

Compute the 1-D metric terms from the element grid arrays x. All the arrays are preallocated by the user and the (square) derivative matrix D should be consistent with the reference grid r used in creategrid!.

If Nq = size(D, 1) and nelem = div(length(x), Nq) then the volume arrays x, J, and ξx should all have length Nq * nelem. Similarly, the face arrays sJ and nx should be of length nface * nelem with nface = 2.

source

computemetric!(x, y, J, ξx, ηx, ξy, ηy, sJ, nx, ny, D)

Compute the 2-D metric terms from the element grid arrays x and y. All the arrays are preallocated by the user and the (square) derivative matrix D should be consistent with the reference grid r used in creategrid!.

If Nq = size(D, 1) and nelem = div(length(x), Nq^2) then the volume arrays x, y, J, ξx, ηx, ξy, and ηy should all be of size (Nq, Nq, nelem). Similarly, the face arrays sJ, nx, and ny should be of size (Nq, nface, nelem) with nface = 4.

source

computemetric!(x, y, z, J, ξx, ηx, ζx, ξy, ηy, ζy, ξz, ηz, ζz, sJ, nx,
               ny, nz, D)

Compute the 3-D metric terms from the element grid arrays x, y, and z. All the arrays are preallocated by the user and the (square) derivative matrix D should be consistent with the reference grid r used in creategrid!.

If Nq = size(D, 1) and nelem = div(length(x), Nq^3) then the volume arrays x, y, z, J, ξx, ηx, ζx, ξy, ηy, ζy, ξz, ηz, and ζz should all be of length Nq^3 * nelem. Similarly, the face arrays sJ, nx, ny, and nz should be of size Nq^2 * nface * nelem with nface = 6.

The curl invariant formulation of Kopriva (2006), equation 37, is used.

Reference: Kopriva, David A. "Metric identities and the discontinuous spectral element method on curvilinear meshes." Journal of Scientific Computing 26.3 (2006): 301-327. https://doi.org/10.1007/s10915-005-9070-8

source